requirements.txt
discord.py
requests
feedparser
deep-translator